Rolled back mermaid dependency to 9.3.4
Mermaid introduced a change to the render API, so a rollback to a working version (9.3.4) was required. The more recent versions change from mermaid.mermaidAPI.render to mermaid.render. If we want to bump the version, we'll need to change the line in Mermaid.svelte that renders the graph.

Testing Done: manually tested that the graph appears, auto-updates when changes are made to the yaml, and intellisense works as expected.